+++ src/components/tsne/TsneSubscriber.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import PubSub from 'pubsub-js';
-import { CELLS_ADD, CELLS_SELECTION, STATUS_INFO } from '../../events';
+import { CELLS_ADD, CELLS_SELECTION, STATUS_INFO, SELECTION_MODE_SET } from '../../events';
 import Tsne from './Tsne';
+import AbstractSelectionSubscriberComponent from '../AbstractSelectionSubscriberComponent'
 
-export class TsneSubscriber extends React.Component {
+
+export class TsneSubscriber extends AbstractSelectionSubscriberComponent {
   constructor(props) {
     super(props);
     this.state = {cells: {}, selectedCellIds: {}};
@@ -11,12 +13,16 @@ export class TsneSubscriber extends React.Component {
 
   componentWillMount() {
     this.cellsAddToken = PubSub.subscribe(CELLS_ADD, this.cellsAddSubscriber.bind(this));
+
     this.cellsSelectionToken = PubSub.subscribe(CELLS_SELECTION, this.cellsSelectionSubscriber.bind(this));
+    this.selectionModeSetToken = PubSub.subscribe(SELECTION_MODE_SET, this.selectionModeSetSubscriber.bind(this));
   }
 
   componentWillUnmount() {
     PubSub.unsubscribe(this.cellsAddToken);
+
     PubSub.unsubscribe(this.cellsSelectionToken);
+    PubSub.unsubscribe(this.selectionModeSetToken);
   }
 
   cellsAddSubscriber(msg, cells) {
@@ -32,6 +38,7 @@ export class TsneSubscriber extends React.Component {
       <Tsne
         cells={this.state.cells}
         selectedCellIds={this.state.selectedCellIds}
+        isRectangleSelection={this.state.isRectangleSelection}
         updateStatus={(message) => PubSub.publish(STATUS_INFO, message)}
         updateCellsSelection={(selectedCellIds) => PubSub.publish(CELLS_SELECTION, selectedCellIds)}
       />